Move domain events command to a better place 👼

+++ src/Shared/Infrastructure/Bus/Event/MySql/MySqlDoctrineDomainEventsConsumer.php
@@ -33,7 +33,9 @@ final class MySqlDoctrineDomainEventsConsumer
 
         $ids = implode(', ', map($this->idExtractor(), $events));
 
-        $this->connection->executeUpdate("DELETE FROM domain_events WHERE id IN ($ids)");
+        if (!empty($ids)) {
+            $this->connection->executeUpdate("DELETE FROM domain_events WHERE id IN ($ids)");
+        }
     }
 
     private function executeSubscribers(callable $subscribers): callable
